+++ src/main/java/revolut/interview/database/dao/TransferDao.java
@@ -3,11 +3,10 @@ package revolut.interview.database.dao;
 import revolut.interview.database.entity.Transfer;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.List;
 
 public interface TransferDao {
-    void doTransfer(BigInteger from, BigInteger to, BigDecimal amount);
+    void doTransfer(Long from, Long to, BigDecimal amount);
     List<Transfer> getTransfers();
-    Transfer getTransfer(BigInteger id);
+    Transfer getTransfer(Long id);
 }
