+++ src/main/java/revolut/interview/exception/AccountNotFoundRequestException.java
@@ -1,9 +1,7 @@
 package revolut.interview.exception;
 
-import java.math.BigInteger;
-
 public class AccountNotFoundRequestException extends RuntimeException {
-    public AccountNotFoundRequestException(BigInteger accountId) {
+    public AccountNotFoundRequestException(Long accountId) {
         super(String.format("Account with id=%s not found", accountId));
     }
 }
